refactor: svg 빌드 크기 제한 문제 해결
